Include project page and also prevent infinte loop
The description field is also shown in the message column on the project
page. This change will turn that text into HTML as well. A new function
convertTextToHtml is created to handle the conversion logic.

This change also tries to solve the infinite loop problem when the
description contains valid less than (<) symbol by checking if the
inner text is rendered differently. This should work!